Added a fuzz target for baggage parsing
